Update README.md
A SequenceVille is a  program that allows users to visualize how different sorting algorithms work. Sorting algorithms are used to rearrange a list of items into a specific order, such as numerical or alphabetical order.

The visualizer typically displays the unsorted list in a form of a bar. The user can then choose a sorting algorithm from a list of options, and the program will show step-by-step how the algorithm sorts the list.

As the algorithm runs, the visualizer highlights the current elements being compared or swapped, and updates the display in real-time to show the progress of the sorting. This can be useful for understanding how different algorithms work and comparing their performance.

SequenceVille also allow users to adjust the speed of the animation or the size of the list being sorted, which can help to demonstrate the impact of different factors on the sorting process